requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
pytest>=9.0.0
pytest-cov>=6.0.0
pydantic>=2.11.0
pydantic-settings>=2.8.0
sqlalchemy>=2.0.43
alembic>=1.14.0
psycopg[binary]==3.2.10
redis==5.0.1
celery==5.3.4
beautifulsoup4==4.12.2
requests==2.31.0
lxml>=5.3.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
stripe==7.4.0
boto3==1.34.34
httpx==0.25.2
aiofiles==23.2.1
pydantic-extra-types>=2.10.0
email-validator==2.1.0
tldextract==5.1.1
langdetect==1.0.9
rake-nltk==1.0.6
nltk==3.8.1
locust==2.15.1
opentelemetry-distro==0.45b0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-sqlalchemy==0.45b0
opentelemetry-instrumentation-redis==0.45b0
opentelemetry-instrumentation-httpx==0.45b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
sentry-sdk[fastapi]==1.38.0
py-spy==0.3.14
memory-profiler==0.61.0